// src/types/house-design.ts
export type HouseDesign = {
  slug: string;
  name: string;
  style: string;
  location: string;
  landArea: number;
  buildingArea: number;
  bedrooms: number;
  floors: number;
  estimatedCost: number;
  estimatedDuration: number;
  description: string;
  highlights: string[];
  image: string;
};